Next iteration: much more extensive format which...
uses the new actor_object lists for source and target. The actors in the first 36 records are complete now; the remainder are partially complete. I'm doing a fair amount of manual editing of the markup (which was generated from the LaTeX in the CAMEO manual) to try to get things closer to what we want the system to look like, not what CAMEO coders at Kansas in 2002 made it look like.